fix endereco crud

+++ backend/plantao-ja-hospital-service/src/main/java/br/com/plantao/plantao_ja/model/endereco/Endereco.java
@@ -21,7 +21,7 @@ public class Endereco {
     @NonNull
     private String rua;
 
-    private Long numero;
+    private int numero;
 
     private String complemento;
 }
